Moved defer and PropertySetMixin to glue.utils.misc
