+++ glue/core/util.py
@@ -2,15 +2,14 @@ from __future__ import absolute_import, division, print_function
 
 import logging
 from itertools import count
-from contextlib import contextmanager
 
 import numpy as np
 import pandas as pd
 
 
 __all__ = ["relim", "split_component_view", "join_component_view",
-           "facet_subsets", "colorize_subsets", "defer", "disambiguate",
-           "row_lookup", "PropertySetMixin"]
+           "facet_subsets", "colorize_subsets", "disambiguate",
+           "row_lookup"]
 
 
 
@@ -192,73 +191,6 @@ def colorize_subsets(subsets, cmap, lo=0, hi=1):
         g = int(255 * g)
         b = int(255 * b)
         subset.style.color = '#%2.2x%2.2x%2.2x' % (r, g, b)
-
-
-class PropertySetMixin(object):
-
-    """An object that provides a set of properties that
-    are meant to encapsulate state information
-
-    This class exposes a properties attribute, which is a dict
-    of all properties. Similarly, assigning to the properties dict
-    will update the individual properties
-    """
-    _property_set = []
-
-    @property
-    def properties(self):
-        """ A dict mapping property names to values """
-        return dict((p, getattr(self, p)) for p in self._property_set)
-
-    @properties.setter
-    def properties(self, value):
-        """ Update the properties with a new dict.
-
-        Keys in the new dict must be valid property names defined in
-        the _property_set class level attribute"""
-        invalid = set(value.keys()) - set(self._property_set)
-        if invalid:
-            raise ValueError("Invalid property values: %s" % invalid)
-
-        for k in self._property_set:
-            if k not in value:
-                continue
-            setattr(self, k, value[k])
-
-
-@contextmanager
-def defer(instance, method):
-    """
-    Defer the calling of a method inside a context manager,
-    and then call it 0 or 1 times afterwards.
-
-    :param instance: The instance of the method to defer
-    :param method: The name of the method to defer
-    :type method: str
-
-    Within the context block, calls to the method will be
-    intercepted, logged, and skipped.
-
-    Upon exiting the context block, the method will be
-    invoked a single time, with the arguments of the
-    most recent invokation inside the context block.
-
-    If the method is never invoked in the context block,
-    it is not called when leaving that block.
-    """
-    history = []
-
-    def log(*a, **k):
-        history.append((a, k))
-
-    orig = getattr(instance, method)
-    setattr(instance, method, log)
-    try:
-        yield
-    finally:
-        setattr(instance, method, orig)
-        for a, k in history[-1:]:
-            orig(*a, **k)
 
 
 def disambiguate(label, taken):
